+++ src/main/java/controller/LadderGameController.java
@@ -1,10 +1,10 @@
 package controller;
 
-import domain.Height;
+import domain.ladder.Height;
 import domain.LadderGame;
-import domain.Player;
-import domain.PlayerNames;
-import domain.Players;
+import domain.player.Player;
+import domain.player.PlayerNames;
+import domain.player.Players;
 import domain.ResultContents;
 import domain.ladder.Ladder;
 import domain.ladder.strategy.GenerateBridgeStrategy;
