+++ src/main/java/domain/ladder/Direction.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public enum Direction {
