+++ src/main/java/domain/ladder/Ladder.java
@@ -1,9 +1,5 @@
 package domain.ladder;
 
-import domain.Direction;
-import domain.Height;
-import domain.Line;
-import domain.Point;
 import domain.ladder.strategy.GenerateBridgeStrategy;
 import java.util.ArrayList;
 import java.util.List;
