+++ src/test/java/domain/player/PlayerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
